Makefile 수정 및 필요없는 인수를 삭제

+++ ft_utils.c
@@ -6,9 +6,9 @@ void	error_msg(void)
 	exit(0);
 }
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	while (s[i])
